Move `DoubleComparison` into `Operators` lint pass
